move plugins/Plugin.js to core/Plugin.js
to better separate concerns and later split into packages
