+++ src/core/Plugin.js
